pyproject.toml
[project]
name = "family-sun"
version = "0.1.0"
authors = [
  { name="Ana Calatayud", email="ana.calatayud@outlook.com" },
]
description = "Python package that generates a sunburst chart from a GEDOM file."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles",
    "fastapi",
    "numpy",
    "pandas",
    "plotly",
    "pre-commit",
    "python-gedcom",
    "python-multipart",
    "ruff",
    "uvicorn"
]
license = "GPL-3.0-or-later"
license-files = ["COPYING"]

[project.urls]
Homepage = "https://github.com/anacltd/family-sun"
Issues = "https://github.com/anacltd/family-sun/issues"

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project.scripts]
generate-tree = "family_sun.generate_family_tree:generate_sunburst_from_gedcom"
